+++ packages/partitura_core/lib/partitura_core.dart
@@ -59,6 +59,7 @@ export 'src/theory/key.dart';
 export 'src/theory/key_signature.dart';
 export 'src/theory/pitch.dart';
 export 'src/theory/scale.dart';
+export 'src/theory/set_theory.dart';
 export 'src/theory/time_signature.dart';
 export 'src/theory/transposition.dart';
 export 'src/theory/triad.dart';